Partial workaround for the Cygwin-specific issue described at https://trac.sagemath.org/ticket/27214#comment:11
Includes regression test.

As noted in the comment, this can still invoke undesired behavior in the
case where an unintialized region of an mmap created with MAP_NORESERVE
is accessed from within a signal handler.

At the time there is no good way to handle that case without extra
support from Cygwin's API which is currently not available as far as I
can tell.  A workaround, for the rare code where this might be
necessary, is to call mprotect (with the appropriate flags) on the
region of memory that needs to be accessed before attempting to access
it.