Replace custom test runner with pytest.
Not to fix what ain't broke, but pytest has evolved quite a bit over
the years, and in particular with pytest-cython it is quite feasible
to run the tests with pytest.

This itself needs more testing, including how it handles failures, and
on other platforms than Linux.

This emulates as much of the funtionality of the original test runner
as possible.  It uses pytest-timeout to set a timeout on tests (this
timeout applies to each individual test though, IIUC, instead of entire
files).  It also runs each test in a separate forked process (again
per-test rather than per-file).

I tried to implement a pytest_runtestloop hook for launching
subprocesses for each test file, and this mostly worked.  But it was
unclear to me from the pytest documentation exactly how test reporting
is supposed to work, so it was hard to figure out how to return test
reports back from the subprocesses.  As it is, the existing --forked
option seems to work well-enough.